Add support to specify prometheus host; doc updates
When running in a container `gethostname` returns the container id,
which isn't very useful in prometheus.  Instead we are likely to want
the real hostname on which we're running, so we can now configure this.

Refactored environment vars out of docker-compose and into a file.
Refactored parameter documentation to be within the sample config.